requirements.txt
# requirements.txt
django==4.2.10
djangorestframework==3.14.0
python-dotenv==1.0.0
openai==1.2.0
pandas==2.1.0
gunicorn==21.2.0
whitenoise==6.5.0